requirements.txt
flask==2.3.3
supabase>=2.0.0
python-dotenv==1.0.0
groq>=0.8.0
together>=1.0.0
pdfplumber==0.7.6
PyPDF2==3.0.1
